requirements.txt
google-cloud-error-reporting==0.30.1
google-cloud-kms==1.0.0
google-cloud-pubsub==0.40.0
google-cloud-storage==1.16.1
google-cloud==0.34.0
html-sanitizer==1.7.2
#html2markdown==0.1.7
flask==1.0.2
psycopg2==2.8.2
#pyjwt[crypto]==1.7.1
pytz==2019.1
PyYAML==5.1
requests==2.21.0
#SQLAlchemy==1.3.4
